requirements.txt
fastapi
python-jose
requests